Refactor: Remove DurationComponent and update TagsComponent
Added corruption tag to indicate at least once corrupted cards.
