+++ app/src/main/java/com/sq/thed_ck_licker/ecs/components/TagsComponent.kt
@@ -1,0 +1,19 @@
+package com.sq.thed_ck_licker.ecs.components
+
+
+enum class CardTag { CARD, MERCHANT, CORRUPTED }
+
+data class TagsComponent(private val tags: List<CardTag> = emptyList()) {
+    fun getTags(): List<CardTag> {
+        return this.tags
+    }
+
+    fun cardIsMerchant(): Boolean {
+        return this.tags.contains(CardTag.MERCHANT)
+    }
+
+    fun addTag(tag: CardTag) {
+        this.tags.plus(tag)
+    }
+
+}
